Improve IR for GF's towards eliminating builtins
This commit contains a LOT of changes.
I'm working on eliminating builtins.
Along the way I had to improve how stamps are read and used
by the generic function discriminating functions.
Currently the interpreter works but the discriminating function
compiler will not work.  It will need to be brought up to date
using the interpreter as a guide.